+++ src/outputs.rs
@@ -8,13 +8,13 @@ pub struct CreateOrderResponse {
 
 #[derive(Debug, Serialize, Deserialize, SchemaWrite, SchemaRead)]
 pub struct DeleteOrderResponse {
-    pub filled_qty: u32,
-    pub average_price: u32,
+    pub filled_qty: u64,
+    pub average_price: u64,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize, SchemaWrite, SchemaRead)]
 pub struct Depth {
-    pub bids: Vec<[u32; 2]>,
-    pub asks: Vec<[u32; 2]>,
+    pub bids: Vec<[u64; 2]>,
+    pub asks: Vec<[u64; 2]>,
     pub last_update_id: String,
 }
